V2.09 Issue #71 + Issue #72 + Issue #73 Fehlerbehebung und Diverses
- Issue #71 Anderer cast für Variable um Fehler zu vermeiden
- Issue #72 Refresh für die Displayanzeige war falsch in der IF-Schleife eingefügt
- Issue #73 Fehlende Datei hochgeladen und überflüssige Dateien gelöscht
- Text in Anzeige für "PVI-Down" und "LM-Down" ersetzt duch Grafiken für "PV Off" und "LM Off"

+++ Frame/DrawImage.h
@@ -218,7 +218,7 @@ int DrawImage(char* fileName, int posx, int posy)
               fbfd,
               0);
 
-    if ((int)fbpI == -1) {
+    if ((intptr_t)fbpI == -1) {
         //printf("Failed to mmap.\n");
     }
     else {
